LayoutTests/platform/ios-sim/fast/canvas/canvas-fillPath-shadow-expected.txt

Ensure correct behavior of canvas with path fill shadow

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS data[0] is 255
PASS data[1] is 20
PASS data[2] is 0
PASS data[0] is 255
PASS data[1] is 20
PASS data[2] is 0
PASS data[0] is 255
FAIL data[1] should be 20. Was 17.
PASS data[2] is 0
PASS data[0] is 255
PASS data[1] is 20
PASS data[2] is 0
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] should not be 255 and it's not.
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] should not be 255 and it's not.
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] should not be 255 and it's not.
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] should not be 255 and it's not.
PASS successfullyParsed is true

TEST COMPLETE
